Revert "changed to long double"
This reverts commit fc669b1f1c584699f20a5bb9db3779b9d6eb5825.

+++ boardToVector.c
@@ -6,7 +6,7 @@
 #include <malloc.h>
 #include "boardstate.h"
 #include "fenToBoardState.h"
-void boardToVector(BSTATE *board, long double *vector)
+void boardToVector(BSTATE *board, double *vector)
 {
   vector[0] = board->WKCFlag;
   vector[1] = board->WQCFlag;
